LU-8324 hsm: prioritize HSM requests
This patch separates the coordinator's implementation from the way
it processes requests stored in its llog.

A new scan_policy is defined: "priority_policy"
This allows to define which types of requests are to be processed
first by the coordinator.
For example it seems a good idea to have RESTORE requests processed
before ARCHIVE requests (unless the fs is full) as a RESTORE is most
certainly issued by a human client.

The working principles of the policy can be found in the comments of
"mdt_coordinator_priority_policy.c".

As allowed by the interface (introduced by this patch), the priority
policy adds two proc files under the hsm proc directory:
- csp_priority_mask:  NOOP ARCHIVE [RESTORE] [REMOVE] [CANCEL]
- csp_priority_ratio: 80

(default values)

The priority_mask defines which requests have which level of
priority. Currently the policy defines two levels of priority:
high and low. The policy also defines the proportion of high and
low priority requests that the coordinator will process each time
it iterates over its llog. That is to say that if ratio is set to 80,
and cdt_max_requests is set to 10, the coordinator will select 8 high
priority requests and 2 low priority requests from the llog (given
that there are enough requests of both level of priority).

Both proc files are writable. Example:
$> echo "+archive -remove" > csp_priority_mask
$> echo "75" > csp_priority_ratio # Defines the high value
But those properties can also be set properly with lctl:
$> lctl set_param mdt.<mdt_name>.hsm.csp_priority_mask=+archive
-remove
$> lctl set_param mdt.<mdt_name>.hsm.csp_priority_ratio=50

Scan policies can be dynamically switched by writing the name of
the policy to use in the proc file "scan_policy" (which can be found
in the hsm proc directory).

This patch also adds tests in sanity-hsm (252, 253, 254 and 255).

+++ lustre/mdt/mdt_hsm_cdt_client.c
@@ -253,7 +253,7 @@ hsm_action_permission(struct mdt_thread_info *mti,
 	struct coordinator *cdt = &mti->mti_mdt->mdt_coordinator;
 	struct lu_ucred *uc = mdt_ucred(mti);
 	struct md_attr *ma = &mti->mti_attr;
-	const __u64 *mask;
+	const int *mask;
 	int rc;
 	ENTRY;
